Add new command to mount with store and router
