+++ tsconfig.json
@@ -26,6 +26,7 @@
   },
   "include": [
     "src",
-    "node_modules/cypress/types/cypress-global-vars.d.ts"
+    "node_modules/cypress/types/cypress-global-vars.d.ts",
+    "cypress"
   ]
 }
